requirements.txt
# PaddleOCR-VL Gradio Application Requirements

# Core dependencies
torch>=2.0.0
torchvision
transformers>=4.40.0
gradio>=5.0.0
pillow>=10.0.0
einops
sentencepiece
protobuf

# Utilities
accelerate>=0.26.0
